requirements.txt
django~=4.2.4
django-environ
django-crontab
djangorestframework
environ
psycopg